refactor: files to atomic structure
