+++ src/components/UI/organism/team.tsx
@@ -1,10 +1,10 @@
-import member1 from "../assets/team-member-1.png"
-import member2 from "../assets/team-member-2.png"
-import member3 from "../assets/team-member-3.png"
-import member4 from "../assets/team-member-4.png"
-import member5 from "../assets/team-member-5.png"
-import member6 from "../assets/team-member-6.png"
-import bg from "../assets/bg-team.svg"
+import member1 from "../../../assets/team-member-1.png"
+import member2 from "../../../assets/team-member-2.png"
+import member3 from "../../../assets/team-member-3.png"
+import member4 from "../../../assets/team-member-4.png"
+import member5 from "../../../assets/team-member-5.png"
+import member6 from "../../../assets/team-member-6.png"
+import bg from "../../../assets/bg-team.svg"
 
 function Team() {
     return (
