--- src/components/UI/pages/main-page.tsx
@@ -1,33 +1,0 @@
-import Header from "../../header";
-import Cover from "../../cover";
-import Catalog from "../../catalog";
-import AboutBlock from "../organism/about-block";
-import Quiz from "../../quiz";
-import Team from "../../team";
-import FAQ from "../../faq";
-import Footer from "../../footer";
-import {categoriesAPI} from "../../../services/CategoriesService";
-import Loader from "../atoms/loader";
-
-
-function MainPage() {
-    const {isLoading} = categoriesAPI.useFetchAllCategoriesQuery('');
-
-    return (
-        <>
-            {isLoading && <Loader />}
-            <Header type={"primary"} />
-            <main className="main">
-                <Cover />
-                <Catalog />
-                <AboutBlock />
-                <Quiz />
-                <Team />
-                <FAQ />
-                <Footer />
-            </main>
-        </>
-    )
-}
-
-export default MainPage
