+++ src/components/pages/main-page.tsx
@@ -1,0 +1,39 @@
+import Header from "../UI/organism/header";
+import Cover from "../UI/molecules/cover";
+import AboutBlock from "../UI/organism/about-block";
+import Quiz from "../UI/organism/quiz";
+import Team from "../UI/organism/team";
+import FAQ from "../UI/organism/faq";
+import Footer from "../UI/organism/footer";
+import Loader from "../UI/atoms/loader";
+import RadiosBlock from "../UI/organism/radios-block";
+import CardsBlock from "../UI/organism/cards-block";
+import {categoriesAPI} from "../../services/CategoriesService";
+
+function MainPage() {
+    const {isLoading} = categoriesAPI.useFetchAllCategoriesQuery('');
+
+    return (
+        <>
+            {isLoading && <Loader />}
+            <Header type={"primary"} />
+            <main className="main">
+                <Cover />
+                <section className="block" id={"catalog"}>
+                    <h2 className="block__title">Catalog</h2>
+                    <div className="block__flex-container">
+                        <RadiosBlock />
+                        <CardsBlock />
+                    </div>
+                </section>
+                <AboutBlock />
+                <Quiz />
+                <Team />
+                <FAQ />
+                <Footer />
+            </main>
+        </>
+    )
+}
+
+export default MainPage
